refacotor: Move uninstall
