+++ src/cargo/ops/mod.rs
@@ -6,7 +6,7 @@ pub use self::cargo_fetch::{fetch, FetchOptions};
 pub use self::cargo_generate_lockfile::generate_lockfile;
 pub use self::cargo_generate_lockfile::update_lockfile;
 pub use self::cargo_generate_lockfile::UpdateOptions;
-pub use self::cargo_install::{install, install_list, uninstall};
+pub use self::cargo_install::{install, install_list};
 pub use self::cargo_new::{init, new, NewOptions, VersionControl};
 pub use self::cargo_output_metadata::{output_metadata, ExportInfo, OutputMetadataOptions};
 pub use self::cargo_package::{package, PackageOpts};
@@ -14,6 +14,7 @@ pub use self::cargo_pkgid::pkgid;
 pub use self::cargo_read_manifest::{read_package, read_packages};
 pub use self::cargo_run::run;
 pub use self::cargo_test::{run_benches, run_tests, TestOptions};
+pub use self::cargo_uninstall::uninstall;
 pub use self::fix::{fix, fix_maybe_exec_rustc, FixOptions};
 pub use self::lockfile::{load_pkg_lockfile, write_pkg_lockfile};
 pub use self::registry::HttpTimeout;
@@ -39,6 +40,7 @@ mod cargo_pkgid;
 mod cargo_read_manifest;
 mod cargo_run;
 mod cargo_test;
+mod cargo_uninstall;
 mod fix;
 mod lockfile;
 mod registry;
